move handlers to separate module
